requirements.txt
django==5.0.4
dj-database-url
dj-static
python-decouple
django-environ
gunicorn
psycopg2-binary
whitenoise
djangorestframework
